requirements.txt
psycopg2
python-dotenv
pandas
requests